requirements.txt
minio==7.1.8
numpy==1.22.3
opencv-python==4.5.5.64
pandas==1.4.2
pymongo==4.1.1
python-dotenv==0.20.0
tensorflow==2.9.1
urllib3==1.26.9
yolov4==3.2.0
